used long type

+++ src/Test.java
@@ -1,6 +1,6 @@
 public class Test {
     public static void main(String[] args) {
-        int a = 10;
+        long a = 10;
         int b;
         b = 15;
         System.out.println(b%a);
